Use SearchMultiSelectInputWithChip on OrganizationMultiSelectInput
- Rename ChipMultiSelectInput to SearchMultiSelectInputWithChip
- Use OrganizationMultiSelectInput with chip prop on EntryForm
